neurocalm-backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.1
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
bcrypt==5.0.0
python-multipart==0.0.20
python-dotenv==1.0.1
pydantic==2.10.4
email-validator==2.2.0
pydantic-settings==2.7.1
requests>=2.31.0,<3
numpy==1.26.4
scipy==1.15.0
fpdf2==2.8.2
psutil==6.1.1
scikit-learn==1.6.1
joblib==1.4.2
pandas==2.2.3
